{"schema":"decodex/commit/1","summary":"Extract frozen frame snapshot resolution","authority":"manual"}
