+++ docs/reference/host-core-reset.md
@@ -86,8 +86,10 @@ The current native-host Swift split is:
   routing, capture-stream preparation, and mouse passthrough.
 - `CaptureOverlayImageSampler.swift`: CoreGraphics below-overlay capture and display-point sample
   adaptation used by overlay windows, live chrome, and scroll fallback acquisition.
-- `FrozenFrameAuthority.swift`: ScreenCaptureKit frozen-frame stream setup, latch-token
-  resolution, self-capture-safe frame gating, and fresh-frame authority decisions.
+- `FrozenFrameAuthority.swift`: ScreenCaptureKit frozen-frame stream setup, self-capture-safe
+  stream gating, and lifecycle state.
+- `FrozenFrameAuthority+SnapshotResolution.swift`: frozen-frame latch-token resolution, RGB/loupe
+  sampling, self-capture-safe frame gating, and fresh-frame authority decisions.
 - `FrozenFrameStreamOutput.swift`: ScreenCaptureKit stream-output delegate adaptation,
   usable-frame filtering, display-time conversion, and frame-record emission into the authority.
 - `FrozenFrameContentFilterPlanner.swift`: frozen-frame display target planning, shareable-content
